feat: Add support for "underlying" types in the signatures
This is a BREAKING CHANGE in three respects.

1. The minimum version of Go has been lifted from 1.17 to 1.18. This is
   required to support generics.
2. "Naked" nils as arguments are no longer accepted as values; they must
   now be explicitly typed.
3. Integral forms will not be implicitly converted to floats. That is,
   expressions such as F(0) are now invalid, but F(0.0) is valid.

I'm not sure what the appropriate language is for point 2, but the
upshot is that (e.g.) `F(nil)` is now invalid.

The intent with this change is to make it possible for people to declare
custom types such as

```go
type foo string
```

and then later freely call

```go
f := foo("bar")
null.S(f)
```

without explicitly casting `f` to its underlying type.