See if we can make this a v1 package again
Just so we don't need to mess about with having /v5 on the end of the package path. The idea is to take the v5 code and just pretend it is v1, then continue as a v1 package with any changes going forward.

We only want to use this with our own core package, and incompatible version updates aren't really a problem for us. It's public, so there's a tiny chance there are other users, but that seems very unlikely at this point.